refactor: change searcher to global
