+++ binding/rust/bin/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "ip2region"
+default-run = "ip2region"
+version = "0.1.0"
+edition = "2021"
+rust-version = "1.66.0"
+description = "the rust binding for ip2region"
+license = "Apache-2.0"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+search = { path = "../search" }
+clap = { version = "4.0" }
