+++ binding/rust/search/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "search"
+version = "0.1.0"
+edition = "2021"
+rust-version = "1.66.0"
+description = "the rust binding for ip2region"
+license = "Apache-2.0"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+once_cell = "1.16"
+
+[dev-dependencies]
+criterion = "0.3"
+rand = "0.8"
+
+#[[bench]]
+#name = "search"
+#harness = false
