--- binding/rust/src/main.rs
@@ -1,10 +1,0 @@
-mod ip_value;
-mod search;
-
-fn main() {
-    println!("");
-    // let filepath = "../../data/ip2region.xdb";
-    // let searcher = lib::Searcher::new(filepath).expect("load file error");
-    // let result = searcher.search_by_ip("1.2.165.128");
-    // println!("{:?}", result);
-}
